+++ CHANGELOG.md
@@ -8,6 +8,9 @@ Note that since we don't clearly distinguish between a public and private interf
 - [Breaking] Reduce memory usage of `SymmetryOperator.ArrayMapping`
     - Requires calling methods from instance
 - [Breaking] Fix `mol-model/structure/model/properties/seconday-structure.ts` file name (#938)
+- [Breaking] Add `Canvas3DContext` runtime props
+    - pixelScale, pickScale, transparency (blended, wboit, dpoit)
+    - replacing instantiation-time attribs
 - [Breaking] Change default compile target to ES2018
 
 ## [Unreleased]